% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkmode_hook.R
\name{darkmode_hook}
\alias{darkmode_hook}
\title{Chunk hook for darkmode}
\usage{
darkmode_hook(x, options, ...)
}
\arguments{
\item{x}{chunk output}

\item{options}{chunk options}

\item{...}{stuff to pass to default chunk hook}
}
\description{
Chunk hook for darkmode
}
\keyword{internal}
