% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{darkmode_theme_set}
\alias{darkmode_theme_set}
\title{Set the dark and light ggplot themes for darkmode}
\usage{
darkmode_theme_set(
  dark = ggthemes::theme_solarized(light = FALSE),
  light = ggthemes::theme_solarized(light = TRUE)
)
}
\arguments{
\item{dark}{the dark theme}

\item{light}{the light theme}
}
\description{
Set the dark and light ggplot themes for darkmode
}
