% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{knit_print.data.frame}
\alias{knit_print.data.frame}
\title{Print tables in kable or paged formats}
\usage{
\method{knit_print}{data.frame}(x, options, ...)
}
\arguments{
\item{x}{an object with class data.frame}

\item{options}{chunk options}

\item{...}{additional arguments to be passed onto print()}
}
\value{
html for kable or paged tables
}
\description{
Print tables in kable or paged formats
}
