% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{knit_print.ggplot}
\alias{knit_print.ggplot}
\title{Print ggplots in light/dark mode}
\usage{
\method{knit_print}{ggplot}(x, options, ...)
}
\arguments{
\item{x}{an object with class ggplot}

\item{options}{chunk options}

\item{...}{additional arguments to be passed onto print()}
}
\value{
html for light and dark mode plots
}
\description{
Print ggplots in light/dark mode
}
